+++ templates/woocommerce/order-received.php
@@ -1,0 +1,13 @@
+<div id="thegift-order-received">
+  <p>Congratulations! You have taken the first step to better yourself. It is my prayer that you will develop a deeper understanding of yourself, your life and your purpose.</p>
+  <h3>Important Information</h3>
+  <p>Here are a few things to get you started:</p>
+  <ul>
+    <li>The account you created during checkout will be the account you use to log in to access the course material. Click the Get Started button below to log in and go to your account.</li>
+    <li>Your account dashboard will list the courses you are enrolled in as well as a few additional resources that you can utilize.</li>
+    <li>You should have received an email with information about your order. Please check your spam or junk folder if you don't see it right away.</li>
+    <li>Be sure to add website@thegiftccc.com to your safe-sender list. If you need to reset your password, you will get an email from that sender to reset your password.</li>
+    <li>If you have any issues or need any assistance along the way, feel free to email tracy@thegiftccc.com and we'll be happy to help!</li>
+  </ul>
+  <a href="/my-account" class="button">Get Started!</a>
+</div>