+++ Changelog.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## Version 1.1.1 - May 31, 2024
+- Update Styles for Lessons
+- Update WooCommerce Order Received content
+- Remove 'Info' label on Course Overview
+
 ## Version 1.1.0 - May 30, 2024
 - Add support for 'Info' lesson type